requirements.txt
requests
beautifulsoup4
google-generativeai
python-dotenv
protobuf
grpcio
selenium
webdriver-manager 